requirements.txt
google-genai
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
requests>=2.28.0
composio>=0.4.0
